requirements.txt
flask
flask-cors
firebase-admin
google-generativeai
python-dotenv
